Cell sub-types

+++ ChiTech/ChiMath/SpatialDiscretization/FiniteVolume/CellViews/fv_cellbase.h
@@ -6,11 +6,11 @@
 class CellFVValues
 {
 public:
-  int dofs;
+  size_t dofs;
   double                volume=0.0;
   std::vector<double>   face_area = {}; ///< Actually areas
 
-  explicit CellFVValues(int num_dofs) :
+  explicit CellFVValues(size_t num_dofs) :
     dofs(num_dofs),
     face_area({})
     {}
